backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
sqlalchemy
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
requests
